V0.0.1.14【添加】管理后台-菜品管理 Dish 完成
1.使用的纯mybatis，未使用MP，代码比较多。复杂。
2.本来dish菜品与其他模块关联性较强，较为复杂

+++ src/main/java/com/bing/service/CategoryService.java
@@ -1,12 +1,18 @@
 package com.bing.service;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.bing.common.CustomException;
 import com.bing.dao.CategoryDao;
 import com.bing.entity.CategoryDO;
 import com.bing.entity.DTO.CategoryDTO;
+import com.bing.entity.DishDO;
+import com.bing.entity.VO.CategoryVO;
 import com.bing.util.MyBeanUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 /**
  * 菜品及套餐分类(Category)表服务接口
@@ -18,6 +24,12 @@ import org.springframework.stereotype.Service;
 public class CategoryService {
     @Autowired
     private CategoryDao categoryDao;
+
+    @Autowired
+    private DishService dishService;
+
+    @Autowired
+    private SetmealService setmealService;
 
     /**
      * 通过ID查询单条数据
@@ -33,6 +45,21 @@ public class CategoryService {
     }
 
     /**
+     * 通过type 和 ID查询 列表
+     *
+     * @param type 类型
+     * @return 实例对象
+     */
+    public List<CategoryDTO> queryByType(CategoryVO category) {
+        LambdaQueryWrapper<CategoryDO> lambdaQueryWrapper = new LambdaQueryWrapper<>();
+        lambdaQueryWrapper.eq(category.getType() != null, CategoryDO::getType, category.getType());
+        lambdaQueryWrapper.orderByAsc(CategoryDO::getSort).orderByDesc(CategoryDO::getUpdateTime);
+        List<CategoryDO> categoryDO = categoryDao.selectList(lambdaQueryWrapper);
+        List<CategoryDTO> categoryDTO = MyBeanUtil.copyList(categoryDO, CategoryDTO.class, null);
+        return categoryDTO;
+    }
+
+    /**
      * 分页查询
      *
      * @param category    筛选条件
@@ -40,8 +67,10 @@ public class CategoryService {
      * @return 查询结果
      */
     public Page<CategoryDO> queryByPage(CategoryDTO category, int currentPage, int pageSize) {
+        LambdaQueryWrapper<CategoryDO> lambdaQueryWrapper = new LambdaQueryWrapper<>();
+        lambdaQueryWrapper.orderByAsc(CategoryDO::getSort);
         Page<CategoryDO> categoryDOPage = new Page<>(currentPage, pageSize);
-        categoryDao.selectPage(categoryDOPage, null);
+        categoryDao.selectPage(categoryDOPage, lambdaQueryWrapper);
         return categoryDOPage;
     }
 
@@ -60,7 +89,7 @@ public class CategoryService {
     }
 
     /**
-     * 修改数据
+     * 根据 ID 修改数据,
      *
      * @param category 实例对象
      * @return 实例对象
@@ -73,12 +102,29 @@ public class CategoryService {
     }
 
     /**
-     * 通过主键删除数据
+     * 通过主键删除数据，
      *
      * @param id 主键
      * @return 是否成功
      */
     public boolean deleteById(Long id) {
+//        删除之前需要先判断是否关联有菜品，否则不予以删除
+// 根据 分类ID 查询菜品表是否引用此分类
+        DishDO dishDO = new DishDO();
+        dishDO.setCategoryId(id);
+        int countDish = dishService.count(dishDO);
+
+        if (countDish > 0) {
+            throw new CustomException("当前分类下关联了【菜品】，不能删除!");
+        }
+
+//        LambdaQueryWrapper<DishDO> dishLambdaQueryWrapper = new LambdaQueryWrapper<>();
+//        dishLambdaQueryWrapper.eq(DishDO::getCategoryId, id);
+        int countSetmeal = 0;
+        if (countSetmeal > 0) {
+            throw new CustomException("当前分类下关联了【套餐】，不能删除!");
+        }
+
         return categoryDao.deleteByIdSql(id) > 0 ? true : false;
     }
